backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
mediapipe==0.10.7
Pillow==10.0.1